Issue/16 ilnyr

+++ src/main/java/com/itis/service/impl/EventServiceImpl.java
@@ -30,6 +30,12 @@ import static java.util.Objects.isNull;
 @Service
 public class EventServiceImpl implements EventService {
 
+//    private final String[] CSVFormats = {"application/vnd.ms-excel", "text/csv"};
+
+    private List<String> CSVFormats = Arrays.asList(
+            "application/vnd.ms-excel",
+            "text/csv");
+
     private static final Logger LOGGER = Logger.getLogger(EventServiceImpl.class);
 
     private final EventRepository eventRepository;
@@ -59,6 +65,9 @@ public class EventServiceImpl implements EventService {
     @Override
     public void createTimetable(MultipartFile file) {
         try {
+            if (!CSVFormats.contains(file.getContentType())) {
+                throw new IllegalArgumentException("incorrect format of csv file");
+            }
             eventRepository.deleteAll();
             Set<EventParsingForm> eventParsingForms = CSVParser.parse(file.getBytes(), EventParsingForm.class);
 //            Set<EventParsingForm> eventParsingForms = CSVParser.parse(file.getBytes(), EventParsingForm.class);
